kitchen/pyproject.toml
[tool.poetry]
name = "kitchen"
version = "0.1.0"
description = ""
authors = ["Arturo Cuicas <arturo.cuicas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.51.1"
grpcio-tools = "^1.51.1"
grpc-interceptor = "^0.15.0"
python-dotenv = "^1.0.1"
pydantic = "^2.6.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
